Move SecureCheckoutRc to molecules
